iposhala_test/iposhala_test/requirements.txt
pandas
pymongo
python-dotenv
requests
beautifulsoup4
python-dateutil
yfinance
nselib
fastapi
uvicorn
